feat: component rename for better clarity
