+++ src/features/user/Settings/EditProfile/AddressManagment/AddressDeleteModal.tsx
@@ -1,0 +1,11 @@
+import styles from './AddressManagement.module.scss';
+
+const AddressDeleteModal = () => {
+  return (
+    <div className={styles.addrConfirmContainer}>
+      <></>
+    </div>
+  );
+};
+
+export default AddressDeleteModal;
